Add command for manipulating lights
Using dataclasses for easier and prettier declaration of the data
heavy message classes. And in the hopes that the dataclass
functionality could be usefull for printing and comparing messages.

the variable in the __post_init__ method could be default arguments to
the auto generated __init__ method of the dataclass, but I feel having
them as default arguments implies that you could want to use a
non-default value at some time which is not the case, since then the
drone would not respond. Therefore the variables are defined after the
class is initialized.