Add method to send auto heading on command dynamically
Johannes wanted to see if adding methods dynamically to the TcpClient based on the
message definitions in drone_commands.py could be done. This is
because we do now want a user of the TcpClient to have to import
drone_commands since they contains implementation specific information about the message
payload and such that the user should not have to care about. This
could also be achieved by writing out methods on the TcpClient such as

def send_auto_heading_on_command()
    self.send_msg(AutoHeadingOnCommand.to_binary)

but the methods to send each message type will be very similar , so
the idea was that the method to send the command could be dynamically generated
from the message class definition thus eliminating the double record
keeping. The messages without any datafields
have been made static such that they can be sent without first
creating a message object. This is why to_binary has been made into a
class method instead of a property.