Add methods for auto depth and auto heading on/off dynamically
Using partial is needed because python looks up the definition of the
methods added with setattr  when executed, and not when they were defined. When
add_commands is finished executing the 'commands' variable in the for
loop will hold the value of the final for loop iteration. Thus all the
methods added in the for loop will send the last command from the last
for loop iteration, see:
https://stackoverflow.com/questions/12423614/local-variables-in-nested-functions
using a partial fixes this