Add command dynamically for setting lights
The docstring for the SetLightCommand and the actual class do not tell
the same story since the parameters for the data fields of the light
command have to be passed as *args to the to_binary class method, but
the input arguments to the TcpClient.set_light_command() still have to
be documented somehow. This is a issue with auto generating the
methods.