Generate send command methods from json protocol definition
A generator script will generate the methods for sending the commands
specified in the json protocol using the jinja2 templating
framework.
If we ensure the generator script is always run upon protocol change
the python methods for sending the commands will not have to be
updated manually. This is less work and removes the possibility of
human error when updating the python methods in response to a protocol
change. It will also enable the c++ client to be based on the same
json definition which could ensure they are always in sync.
We have moved away from generating the methods dynamically and instead
generate the methods in a python file. This is because it is easier to
use and document.

The json protocol defined here is a WIP and by no means final