Choose protocol version dynamically when instantiating a TcpClient
TcpClient is now a simple factory for choosing the protocol version to
use. The inheritance is flipped around, i.e The protocol versions
inherits from the TcpClientBase instead of the other way around, this
is because the TcpClientBase is the thing that is in common for all
protocol versions, so it makes sense that the protocol versions
inherit from and extend the TcpBaseClient with a version specific
command set.

Co-authored-by: Sindre Hansen <sindrehansen92@gmail.com>