Make TcpClient thread deamonic, and let python handle cleanup
when the thread is non deamonic it would not exit cleanly on exit. Since the
pinging thread is only useful when the programs main thread is running
it is fine for it to exit once the main thread exits. Also removed the
__del__ method since you are not advised to use it and we cannot
guarantee that is called, and there is no need to close the socket
when we are reusing addresses.

Co-authored-by: Sindre Hansen <sindrehansen92@gmail.com>