Move example use to docstring, format example as a doctest
If we add continuous  integration for running doctest it would ensure all
examples run and are up to date. The example shows use of TcpClient
even though it is a docstring for TcpClientBase, this is because we
do not want the user to use the TcpClientBase directly, but rather the
TcpClient

Changed the logging level in the example to warning to avoid the
watchdog messages being printed, as it makes defining the expected
output for doctest harder.